requirements.txt
azure-storage-blob
fastapi
azure-identity
azure-keyvault-secrets
langchain
langchain_openai
langchain-community
pydantic
uvicorn
chromadb